feat: move spinner icon
