+++ packages/react/src/icons/Spinner.tsx
